requirements.txt
annotated-types==0.6.0
anyio==3.7.1
bcrypt==4.0.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.5
defusedxml==0.7.1
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
exceptiongroup==1.1.3
fastapi==0.104.1
h11==0.14.0
httpie==3.2.2
idna==3.4
logging==0.4.9.6
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.4
passlib==1.7.4
pyasn1==0.5.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.16.1
pymongo==4.5.0
PySocks==1.7.1
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
urllib3==2.1.0
uvicorn==0.23.2
